fix(prod-sample): clarify use case
