--- prod-sample/sample-modules/franks-burgers/0.0.0/src/components/Order.jsx
@@ -1,40 +1,0 @@
-/*
- * Copyright 2020 American Express Travel Related Services Company, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-import React, { Suspense } from 'react';
-import { FormattedMessage } from 'react-intl';
-
-const Burger = React.lazy(() => import(/* webpackChunkName: 'Burger' */ './Burger'));
-
-export default function Order() {
-  const [loadBurger, setBurgerLoad] = React.useState(false);
-
-  return (
-    <section>
-      {loadBurger ? (
-        <Suspense fallback={<p><FormattedMessage id="loading" /></p>}>
-          <Burger />
-        </Suspense>
-      ) : null}
-
-      <footer>
-        <button id="order-burger-btn" type="button" onClick={() => setBurgerLoad((state) => !state)}>
-          <FormattedMessage id="franks-cta" />
-        </button>
-      </footer>
-    </section>
-  );
-}
